+++ eesampling/tests/test_diagnostics.py
@@ -22,7 +22,6 @@ import pandas as pd
 
 import pytest
 
-from eesampling.diagnostics import Diagnostics
 from eesampling.model import StratifiedSampling
 
 
@@ -39,21 +38,3 @@ def diagnostics_obj(df_treatment, df_pool, col_name):
 def test_equivalence(diagnostics_obj):
     equivalence = diagnostics_obj.equivalence()
     assert equivalence["ks_ok"].all() == True and equivalence["t_ok"].all() == True
-
-
-def test_record_based_equivalence_euclidean(diagnostics_obj, df_treatment, df_pool, df_equiv):
-
-    
-    equivalence = diagnostics_obj.records_based_equivalence_euclidean(
-        df_equiv, equiv_groupby_col="month", equiv_value_col="baseline_predicted_usage"
-    )
-    assert equivalence
-
-
-def test_record_based_equivalence_chisquare(diagnostics_obj, df_treatment, df_pool, df_equiv):
-
-    
-    equivalence = diagnostics_obj.records_based_equivalence_chisquare(
-        df_equiv, equiv_groupby_col="month", equiv_value_col="baseline_predicted_usage"
-    )
-    assert equivalence